requirements.txt
streamlit
langchain
chromadb
PyMuPDF
python-docx
tiktoken
groq
langchain_community